requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
google-generativeai==0.8.5
gunicorn==23.0.0